refactor: make a file structure, add server implementation based on forest fire
