--- run.py
@@ -1,3 +1,0 @@
-from ecosystem.server import server
-
-server.launch()
